-[x] refactor: code refactor
