+++ public/css/webportal-app.css
@@ -2250,6 +2250,12 @@ video {
 .pb-\[71px\] {
   padding-bottom: 71px;
 }
+.pt-2 {
+  padding-top: 0.5rem;
+}
+.pr-4 {
+  padding-right: 1rem;
+}
 .pt-5 {
   padding-top: 1.25rem;
 }
@@ -2258,9 +2264,6 @@ video {
 }
 .pt-10 {
   padding-top: 2.5rem;
-}
-.pt-2 {
-  padding-top: 0.5rem;
 }
 .pt-11 {
   padding-top: 2.75rem;
@@ -2291,9 +2294,6 @@ video {
 }
 .pt-3 {
   padding-top: 0.75rem;
-}
-.pr-4 {
-  padding-right: 1rem;
 }
 .pt-0\.5 {
   padding-top: 0.125rem;
