+++ public/mix-manifest.json
@@ -1,10 +1,10 @@
 {
-    "/js/app.js": "/js/app.js?id=c3d1c49944d0d4c1fe6ce040f88a89b8",
+    "/js/app.js": "/js/app.js?id=766a15583aec7b4206cbe259d9e38603",
     "/js/script.js": "/js/script.js?id=4920eb85d84da1e87cb71769c04a12a2",
     "/js/webportal-script.js": "/js/webportal-script.js?id=f9c23d4bccd261db9414465886315f45",
     "/js/formbuilder.js": "/js/formbuilder.js?id=151384c2b4e7d700f9822f8de83ea9f5",
     "/manifest.js": "/manifest.js?id=3ed1124c8f4dcb98acc36de67cb80349",
-    "/css/webportal-app.css": "/css/webportal-app.css?id=a6634fa79e4471dfe5b6e5498d186ef2",
-    "/css/app.css": "/css/app.css?id=413fb6e7bd3b0ca011c931ed457ed4cc",
+    "/css/webportal-app.css": "/css/webportal-app.css?id=d8d3d4a1c8204a9c0bc0f1772047c8b9",
+    "/css/app.css": "/css/app.css?id=67580d8ec9f42bb492a4df4da0c526fe",
     "/js/vendor.js": "/js/vendor.js?id=ad893ae286ee59c0659830c39147d414"
 }
